+++ src/components/SignUp/PetPicture.style.tsx
