+++ src/components/SignUp/PetSignUpPicture.tsx
