+++ src/pages/SignUpPages/SignUpProfile.tsx
@@ -4,7 +4,7 @@ import Button from '../../components/Form/Button';
 import Input from '../../components/Form/Input';
 import SelectInput from '../../components/Form/SelectInput';
 import { BIRTH_REGEX } from '../../data/regExp';
-import Form from './SignUp.style';
+import Form from '../SignUp.style';
 
 interface SignUpProfileType {
   userName: string;
@@ -83,5 +83,4 @@ function SignUpProfile() {
 
 export default SignUpProfile;
 
-const StyledSignUpProfile = styled.div`
-`;
+const StyledSignUpProfile = styled.div``;
