+++ src/pages/SignUpPages/SignUpStart.style.tsx
