+++ src/pages/SignUpPages/index.ts
@@ -1,0 +1,6 @@
+import SignEnd from './SignUpEnd';
+import SignPet from './SignUpPet';
+import SignProfile from './SignUpProfile';
+import SignStart from './SignUpStart';
+
+export { SignStart, SignEnd, SignPet, SignProfile };
